requirements.txt
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
config==0.5.1
cryptography==44.0.0
dnspython==2.7.0
flask==3.1.0
flask-jwt-extended==4.7.1
idna==3.10
pymongo==4.10.1
itsdangerous==2.2.0
Jinja2==3.1.5
pyjwt==2.10.1
MarkupSafe==3.0.2
pycparser==2.22
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
urllib3==2.3.0
vercel==0.2.1
Werkzeug==3.1.3
